feat: working on the testlogbase changes
