+++ handshake/services/DBService/models/attachmentBase.py
@@ -1,4 +1,4 @@
-from handshake.services.DBService.models.result_base import SuiteBase
+from handshake.services.DBService.models.result_base import SuiteBase, RunBase
 from tortoise.models import Model
 from tortoise.fields import (
     ForeignKeyField,
@@ -6,9 +6,9 @@ from tortoise.fields import (
     TextField,
     IntField,
     DatetimeField,
+    CharField,
     BooleanField,
     CharEnumField,
-    CharField,
     JSONField,
 )
 from handshake.services.DBService.models.enums import LogType
@@ -53,4 +53,36 @@ class EntityLogBase(Model):
     tags = JSONField(
         description="comma separated list of tags used by the framework to filter the suites or spec files",
         default=[],
+    )
+
+
+class TestLogBase(Model):
+    test: ForeignKeyRelation[RunBase] = ForeignKeyField(
+        "models.RunBase", related_name="log", to_field="testID", pk=True
+    )
+    type = CharEnumField(LogType, description="Log type", null=False)
+    message = TextField(description="Log Message", null=False)
+    generatedBy = CharField(
+        description="which job generated this, Null if generated by the user",
+        null=True,
+        default="NA",
+        max_length=200,
+    )
+    tags = JSONField(
+        description="comma separated list of tags used by the framework to filter the suites or spec files",
+        default=[],
+    )
+    dropped = DatetimeField(auto_now=True, description="timestamp", null=False)
+    apiGenerated = BooleanField(
+        default=False,
+        description="is it generated by api endpoints, if so generatedBy will have api endpoint name",
+    )
+    schedulerGenerated = BooleanField(
+        default=False,
+        description="is it generated by the scheduler jobs, if so the generatedBy will have job Name",
+    )
+
+    userGenerated = BooleanField(
+        default=False,
+        description="is it attached by the user, if so generatedBy will be NULL",
     )
